Move RegisterActivityTest.java to ProblematicTests folder at root
