+++ app/src/androidTest/java/com/sdp/movemeet/MainActivityTest.java
@@ -44,7 +44,6 @@ public class MainActivityTest {
 
     }
 
-
 //    @Test
 //    public void mainActivity_logout() {
 //        Intents.init();
